Improve Codebase
