+++ composer.json
@@ -35,6 +35,7 @@
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.3",
+        "phpstan/phpstan": "^0.9.0",
         "phpunit/phpunit": "^6.0"
     },
     "autoload": {
