+++ src/Components/AbstractComponent.php
@@ -6,13 +6,12 @@
  * @subpackage League\Uri\Components
  * @author     Ignace Nyamagana Butera <nyamsprod@gmail.com>
  * @license    https://github.com/thephpleague/uri-components/blob/master/LICENSE (MIT License)
- * @version    1.4.0
+ * @version    1.5.0
  * @link       https://github.com/thephpleague/uri-components
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-declare(strict_types=1);
 
 namespace League\Uri\Components;
 
@@ -46,7 +45,7 @@ abstract class AbstractComponent implements ComponentInterface
     /**
      * new instance
      *
-     * @param mixed $data the component value
+     * @param string|null $data the component value
      */
     public function __construct(string $data = null)
     {
@@ -107,5 +106,29 @@ abstract class AbstractComponent implements ComponentInterface
         }
 
         return new static($value);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function __debugInfo()
+    {
+        return ['component' => $this->getContent()];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isNull(): bool
+    {
+        return null === $this->getContent();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isEmpty(): bool
+    {
+        return '' == $this->getContent();
     }
 }
