+++ src/Components/AbstractHierarchicalComponent.php
@@ -6,13 +6,12 @@
  * @subpackage League\Uri\Components
  * @author     Ignace Nyamagana Butera <nyamsprod@gmail.com>
  * @license    https://github.com/thephpleague/uri-components/blob/master/LICENSE (MIT License)
- * @version    1.4.0
+ * @version    1.5.0
  * @link       https://github.com/thephpleague/uri-components
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-declare(strict_types=1);
 
 namespace League\Uri\Components;
 
@@ -28,7 +27,7 @@ use IteratorAggregate;
  * @author     Ignace Nyamagana Butera <nyamsprod@gmail.com>
  * @since      1.0.0
  */
-abstract class AbstractHierarchicalComponent implements ComponentInterface, Countable, IteratorAggregate
+abstract class AbstractHierarchicalComponent implements Countable, IteratorAggregate
 {
     use ComponentTrait;
 
@@ -97,14 +96,7 @@ abstract class AbstractHierarchicalComponent implements ComponentInterface, Coun
     /**
      * {@inheritdoc}
      */
-    public function withContent($value): ComponentInterface
-    {
-        if ($value === $this->getContent()) {
-            return $this;
-        }
-
-        return new static($value);
-    }
+    abstract public function withContent($value): ComponentInterface;
 
     /**
      * {@inheritdoc}
