+++ src/Components/UserInfo.php
@@ -6,7 +6,7 @@
  * @subpackage League\Uri\Components
  * @author     Ignace Nyamagana Butera <nyamsprod@gmail.com>
  * @license    https://github.com/thephpleague/uri-components/blob/master/LICENSE (MIT License)
- * @version    1.4.0
+ * @version    1.5.0
  * @link       https://github.com/thephpleague/uri-components
  *
  * For the full copyright and license information, please view the LICENSE
@@ -68,6 +68,22 @@ class UserInfo implements ComponentInterface
             'user' => $this->getUser(),
             'pass' => $this->getPass(),
         ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isNull(): bool
+    {
+        return null === $this->getContent();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isEmpty(): bool
+    {
+        return '' == $this->getContent();
     }
 
     /**
